Edit-highlighting only highlights when appropriate
Essentially, if there is no drawn list that is being edited against,
then no highlighting occurs.

The only remaining issue is that if an empty list is drawn, the user
doesn't see the list but will still get highlights if they edit against
it.  My solution is to draw some sort of "empty list" graphic so that
they understand that a list *has* been processed and drawn, it was just
empty.  I'll implement this at some point.